+++ packages/define-model/src/index.ts
@@ -1,8 +1,14 @@
 import { createUnplugin } from 'unplugin'
-import { createFilter } from '@rollup/pluginutils'
+import { createFilter, normalizePath } from '@rollup/pluginutils'
 import { REGEX_SETUP_SFC, REGEX_VUE_SFC } from '@vue-macros/common'
 import { transformDefineModel } from './core'
-import { emitHelperCode, emitHelperId } from './core/helper'
+import {
+  emitHelperCode,
+  emitHelperId,
+  helperPrefix,
+  useVmodelHelperCode,
+  useVmodelHelperId,
+} from './core/helper'
 import type { FilterPattern } from '@rollup/pluginutils'
 
 export interface Options {
@@ -13,6 +19,12 @@ export interface Options {
    * @default 3
    */
   version?: 2 | 3
+  /**
+   * Unified mode, only works for Vue 2
+   *
+   * Converts `modelValue` to `value`
+   */
+  unified?: boolean
 }
 
 export type OptionsResolved = Omit<Required<Options>, 'exclude'> & {
@@ -23,6 +35,7 @@ function resolveOption(options: Options): OptionsResolved {
   return {
     include: [REGEX_VUE_SFC, REGEX_SETUP_SFC],
     version: 3,
+    unified: true,
     ...options,
   }
 }
@@ -38,15 +51,17 @@ export default createUnplugin((userOptions: Options = {}) => {
     enforce: 'pre',
 
     resolveId(id) {
-      if (id === emitHelperId) return id
+      if (id.startsWith(helperPrefix)) return id
     },
 
     loadInclude(id) {
-      return id === emitHelperId
+      return id.startsWith(helperPrefix)
     },
 
-    load() {
-      return emitHelperCode
+    load(_id) {
+      const id = normalizePath(_id)
+      if (id === emitHelperId) return emitHelperCode
+      else if (id === useVmodelHelperId) return useVmodelHelperCode
     },
 
     transformInclude(id) {
@@ -55,7 +70,7 @@ export default createUnplugin((userOptions: Options = {}) => {
 
     transform(code, id) {
       try {
-        return transformDefineModel(code, id, options.version)
+        return transformDefineModel(code, id, options.version, options.unified)
       } catch (err: unknown) {
         this.error(`${name} ${err}`)
       }