+++ packages/define-model/tests/fixtures/runtime/vue3/with-define-emits.vue
@@ -1,0 +1,12 @@
+<script setup lang="ts">
+const { modelValue } = defineModel<{
+  modelValue: string
+}>()
+
+const emit = defineEmits<{
+  (evt: 'change'): void
+}>()
+
+emit('change')
+modelValue.value = 'hello'
+</script>