+++ playground/vue3/src/App.vue
@@ -1,0 +1,60 @@
+<script setup lang="ts">
+import DefineRenderVue from './examples/define-render/index.vue'
+import DefineOptionsVue from './examples/define-options/index.vue'
+import HoistStaticVue from './examples/hoist-static/index.vue'
+import DefineModelReactivityVue from './examples/define-model/reactivity-transform/parent.vue'
+import DefineModelRuntimeVue from './examples/define-model/runtime/parent.vue'
+
+import { SetupComponentFoo } from './examples/setup-component'
+import { SetupComponentType } from './examples/setup-component/types'
+
+import SetupSFC from './examples/setup-sfc/index.setup'
+import ShortVmodel from './examples/short-vmodel/parent.vue'
+import Full from './examples/full.setup'
+</script>
+
+<template>
+  <fieldset>
+    <legend>defineOptions</legend>
+    <DefineOptionsVue />
+  </fieldset>
+
+  <fieldset>
+    <legend>defineRender</legend>
+    <DefineRenderVue />
+  </fieldset>
+
+  <fieldset>
+    <legend>hostStatic</legend>
+    <HoistStaticVue />
+  </fieldset>
+
+  <fieldset>
+    <legend>defineModel</legend>
+    <DefineModelReactivityVue />
+    <hr />
+    <DefineModelRuntimeVue />
+  </fieldset>
+
+  <fieldset>
+    <legend>setupComponent</legend>
+    <SetupComponentFoo />
+    <hr />
+    <SetupComponentType />
+  </fieldset>
+
+  <fieldset>
+    <legend>setupSFC</legend>
+    <SetupSFC />
+  </fieldset>
+
+  <fieldset>
+    <legend>Short v-model</legend>
+    <short-vmodel />
+  </fieldset>
+
+  <fieldset>
+    <legend>Full example</legend>
+    <Full model-value="123" />
+  </fieldset>
+</template>