+++ packages/define-model/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @vue-macros/define-model
 
+## 0.13.2
+
+### Patch Changes
+
+- [`7877ac7`](https://github.com/sxzz/unplugin-vue-macros/commit/7877ac7d92c0c9b90297364abddae3562236e2d2) Thanks [@sxzz](https://github.com/sxzz)! - fix unifed key
+
+## 0.13.1
+
+### Patch Changes
+
+- [`ac603d0`](https://github.com/sxzz/unplugin-vue-macros/commit/ac603d0018637a615133899940f5ca8bd76b7435) Thanks [@sxzz](https://github.com/sxzz)! - add unified
+
+## 0.13.0
+
+### Minor Changes
+
+- [#104](https://github.com/sxzz/unplugin-vue-macros/pull/104) [`97d82c6`](https://github.com/sxzz/unplugin-vue-macros/commit/97d82c62a3aad1b1158b9697ad6bd9d052c1c086) Thanks [@sxzz](https://github.com/sxzz)! - add runtime defineModel, requires `@vueuse/core`.
+
+  ⚠️ BREAKING CHANGE: original `defineModel` renamed to `$defineModel`.
+
+### Patch Changes
+
+- Updated dependencies [[`97d82c6`](https://github.com/sxzz/unplugin-vue-macros/commit/97d82c62a3aad1b1158b9697ad6bd9d052c1c086)]:
+  - @vue-macros/common@0.12.1
+
 ## 0.12.0
 
 ### Minor Changes