+++ packages/macros/CHANGELOG.md
@@ -1,5 +1,53 @@
 # unplugin-vue-macros
 
+## 0.13.2
+
+### Patch Changes
+
+- Updated dependencies [[`7877ac7`](https://github.com/sxzz/unplugin-vue-macros/commit/7877ac7d92c0c9b90297364abddae3562236e2d2)]:
+  - @vue-macros/define-model@0.13.2
+
+## 0.13.1
+
+### Patch Changes
+
+- [`ac603d0`](https://github.com/sxzz/unplugin-vue-macros/commit/ac603d0018637a615133899940f5ca8bd76b7435) Thanks [@sxzz](https://github.com/sxzz)! - add unified
+
+- [`b00600a`](https://github.com/sxzz/unplugin-vue-macros/commit/b00600a64ec39381613d2b9b069d82e82b1a39e6) Thanks [@sxzz](https://github.com/sxzz)! - add missing dts file
+
+- Updated dependencies [[`ac603d0`](https://github.com/sxzz/unplugin-vue-macros/commit/ac603d0018637a615133899940f5ca8bd76b7435), [`b00600a`](https://github.com/sxzz/unplugin-vue-macros/commit/b00600a64ec39381613d2b9b069d82e82b1a39e6)]:
+  - @vue-macros/define-model@0.13.1
+  - @vue-macros/define-render@0.13.2
+
+## 0.13.0
+
+### Minor Changes
+
+- [#104](https://github.com/sxzz/unplugin-vue-macros/pull/104) [`97d82c6`](https://github.com/sxzz/unplugin-vue-macros/commit/97d82c62a3aad1b1158b9697ad6bd9d052c1c086) Thanks [@sxzz](https://github.com/sxzz)! - add runtime defineModel, requires `@vueuse/core`.
+
+  ⚠️ BREAKING CHANGE: original `defineModel` renamed to `$defineModel`.
+
+### Patch Changes
+
+- Updated dependencies [[`97d82c6`](https://github.com/sxzz/unplugin-vue-macros/commit/97d82c62a3aad1b1158b9697ad6bd9d052c1c086)]:
+  - @vue-macros/define-model@0.13.0
+  - unplugin-vue-define-options@0.12.1
+  - @vue-macros/define-render@0.13.1
+  - @vue-macros/hoist-static@0.12.1
+  - @vue-macros/setup-component@0.12.1
+  - @vue-macros/setup-sfc@0.12.1
+  - @vue-macros/short-emits@0.12.2
+
+## 0.12.1
+
+### Patch Changes
+
+- [#101](https://github.com/sxzz/unplugin-vue-macros/pull/101) [`4153f51`](https://github.com/sxzz/unplugin-vue-macros/commit/4153f51274146c0282b82749a9e04eaefd60414a) Thanks [@sxzz](https://github.com/sxzz)! - add dts for Vue 2.7
+
+- Updated dependencies [[`4153f51`](https://github.com/sxzz/unplugin-vue-macros/commit/4153f51274146c0282b82749a9e04eaefd60414a)]:
+  - @vue-macros/define-render@0.13.0
+  - @vue-macros/short-emits@0.12.1
+
 ## 0.12.0
 
 ### Minor Changes